+++ app/controllers/concerns/authorization.rb
@@ -1,14 +1,20 @@
 module Authorization
   extend ActiveSupport::Concern
 
+  included do
+    helper_method :allowed_to?
+  end
+
   private
 
   def authorize!(record, query)
-    policy = policy_for(record)
-    allowed = policy.public_send(query)
-    raise ApplicationPolicy::NotAuthorizedError unless allowed
+    raise ApplicationPolicy::NotAuthorizedError unless allowed_to?(record, query)
 
     true
+  end
+
+  def allowed_to?(record, query)
+    policy_for(record).public_send(query)
   end
 
   def policy_for(record)