+++ application/config/roles.rb
@@ -1,0 +1,1 @@
+# ApplicationRoles.register :risk_manager, permissions: %w[risks.read risks.manage]